coworkers/goit_web_hw14/pyproject.toml
[tool.poetry]
name = "goit-web-hw14"
version = "0.1.0"
description = ""
authors = ["OlehVakulchyk <4020405@gmail.com>"]
readme = "README.md"
packages = [{include = "goit_web_hw13"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.95.1"
uvicorn = {extras = ["standard"], version = "^0.22.0"}
sqlalchemy = "^2.0.13"
psycopg2 = "^2.9.6"
pydantic = {extras = ["email"], version = "^1.10.7"}
alembic = "^1.10.4"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
libgravatar = "^1.0.4"
faker = "^18.7.0"
fastapi-mail = "^1.2.8"
python-dotenv = "^1.0.0"
redis = "^4.5.5"
fastapi-limiter = "^0.1.5"
git-filter-repo = "^2.38.0"
cloudinary = "^1.33.0"
pytest = "^7.3.1"
pytest-mock = "^3.10.0"
pytest-cov = "^4.1.0"


[tool.poetry.group.dev.dependencies]
sphinx = "^7.0.1"


[tool.poetry.group.test.dependencies]
httpx = "^0.24.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
